+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Death Markers Changelog
 # v1.0.4
 - Added "Enable Respawn Time Override" setting
-- Removed C++ exceptions
+- Added death marker settings button in pause menu (there is an option to turn it off)
+- Added "Enable in Classic Mode" setting
+- Added "Enable in Platformer Mode" setting
+- Added "Enable in Normal Mode" setting
 - Made (and stole) some alternate icons (check the github repo)
+- Removed C++ exceptions
 # v1.0.3
 - New marker icon
 - Fixed death markers showing on both sides in dual/2 plr mode when only one icon crashes ([issue #2](https://github.com/Bean0of/deathmarkers/issues/2))